maps: rmsmatcher: fix display of matches for images with borders
When we display matches over the original texture map image, if the DiscreteMap
generated from the segment data is smaller than the image (which may happen when
the image has borders, depending on how the DiscreteMap was created), the size
reserved for the image in the output and the actual image did not match. This
fixes the problem by using the actual image size when an image is provided.